center the stackview
